fixing a bunch of remaining Codehaus entries
remember: Codehaus was something like Github or Sourceforge in the past,
so the org/codehaus prefix is not the project (we already had that).
Here I retagged a bunch of packages from the Jackson project. Might
still not be the correct package (rather looks like a repackaged thing),
but at least more correct than Codehaus :)